+++ app/app/rules/page.tsx
@@ -1,0 +1,2 @@
+import RuleManagement from "../../rule-management";
+export default function RulesPage(){return <RuleManagement/>;}